+++ backend/src/channel/channel.model.ts
@@ -1,12 +1,23 @@
+import { ApiProperty } from '@nestjs/swagger';
 import { Table, Model, Column } from 'sequelize-typescript';
 import { Notification } from '../notifications/notification.model';
-import { ApiProperty } from '@nestjs/swagger';
+import { User } from '../users/user.model';
 
 @Table
 export class Channel extends Model {
   @ApiProperty()
   @Column
   description: string;
+
+  sendMessage(userName: string, notification: Notification): void {
+    /*
+    TODO: in this part:
+    1. Persist Log
+    2. Use abstract class
+    */
+    const message = `Notification (${notification.category.description}): ${notification.message} for ${userName}. Channel: ${this.description}`;
+    console.log(message);
+  }
 }
 
 export abstract class ChannelDTO {
@@ -15,7 +26,7 @@ export abstract class ChannelDTO {
   abstract sendMessage(notification: Notification): void;
 }
 
-export class SMS extends Channel {
+export class SMS extends ChannelDTO {
   description: string;
 
   sendMessage(notification: Notification): void {
@@ -23,7 +34,7 @@ export class SMS extends Channel {
   }
 }
 
-export class Email extends Channel {
+export class Email extends ChannelDTO {
   description: string;
 
   sendMessage(notification: Notification): void {
@@ -31,7 +42,7 @@ export class Email extends Channel {
   }
 }
 
-export class PushNotification extends Channel {
+export class PushNotification extends ChannelDTO {
   description: string;
 
   sendMessage(notification: Notification): void {